+++ CHANGELOG.markdown
@@ -23,6 +23,10 @@
   -type instance Index Value = Text
   +type instance Index Value = Key
   ```
+* Remove `Primitive` and `AsPrimitive`, since https://tools.ietf.org/html/rfc7159
+  de-emphasized the notion of primitive versus composite JSON values.
+  * `_Number`'s default signature, `_Bool`, `Bool_`, `_String`, `String_`,
+    `_Null`, and `Null_` now have an `AsValue` constraint.
 * Add `Wrapped` and `Rewrapped` instances for `KeyMap`. These treat `KeyMap v`
   as a wrapper around `[(Key, v)]`. The order in which the key-value pairs
   appear in this list is not stable.