+++ CHANGELOG.markdown
@@ -1,3 +1,8 @@
+next [????.??.??]
+-----------------
+* Use `aeson`'s `decode` function in the `AsJSON` instance for `ByteString`s.
+  This allows removing a direct dependency on `attoparsec`.
+
 1.2.2 [2022.08.11]
 ------------------
 * Add `atKey :: AsValue t => Key -> Traversal' t (Maybe Value)`, a variant of