+++ CHANGELOG.markdown
@@ -1,5 +1,5 @@
-next [????.??.??]
------------------
+1.2.3 [2023.06.27]
+------------------
 * Use `aeson`'s `decode` function in the `AsJSON` instance for `ByteString`s.
   This allows removing a direct dependency on `attoparsec`.
 * Allow building with `aeson-2.2`.